+++ branches/dataspaces-epsi/tests/C/test_get_run.c
@@ -38,7 +38,7 @@
 //# of processors in x-y-z direction
 static int npx_, npy_, npz_;
 //block size per processor per direction
-static int spx_, spy_, spz_;
+static uint64_t spx_, spy_, spz_;
 //# of iterations
 static int timestep_;
 //# of processors in the application
@@ -46,7 +46,7 @@ static int npapp_;
 
 static int rank_, nproc_;
 
-static int offx_, offy_, offz_;
+static uint64_t offx_, offy_, offz_;
 
 static struct timer timer_;
 
@@ -62,27 +62,27 @@ Matrix representation
 |
 v (y)
 */
-static double* allocate_2d(int x, int y)
+static double* allocate_2d(uint64_t x, uint64_t y)
 {
     double* tmp = NULL;
     tmp = (double*)malloc(elem_size_ * x * y);
     return tmp;
 }
 
-static void set_offset_2d(int rank, int npx, int npy, int spx, int spy)
+static void set_offset_2d(int rank, int npx, int npy, uint64_t spx, uint64_t spy)
 {
     offx_ = (rank % npx) * spx;
     offy_ = (rank / npx) * spy;
 }
 
-static double* allocate_3d(int x, int y, int z)
+static double* allocate_3d(uint64_t x, uint64_t y, uint64_t z)
 {
     double *tmp = NULL;
     tmp = (double*)malloc(elem_size_* x * y * z);
     return tmp;
 }
 
-static void set_offset_3d(int rank, int npx, int npy, int npz, int spx, int spy, int spz)
+static void set_offset_3d(int rank, int npx, int npy, int npz, uint64_t spx, uint64_t spy, uint64_t spz)
 {
     offx_ = (rank % npx) * spx;
     offy_ = (rank / npx % npy) * spy;
@@ -103,17 +103,17 @@ static int couple_read_2d(unsigned int ts, int num_vars, enum transport_type typ
 	//get data from space
     set_offset_2d(rank_, npx_, npy_, spx_, spy_);
 	int elem_size = elem_size_;
-	int xl = offx_;
-	int yl = offy_;
-	int zl = 0;
-	int xu = offx_ + spx_ - 1;
-	int yu = offy_ + spy_ - 1;
-	int zu = 0;
+	uint64_t xl = offx_;
+	uint64_t yl = offy_;
+	uint64_t zl = 0;
+	uint64_t xu = offx_ + spx_ - 1;
+	uint64_t yu = offy_ + spy_ - 1;
+	uint64_t zu = 0;
 	double tm_st, tm_end, tm_max, tm_diff;
 	int root = 0;
 
 #ifdef DEBUG
-	uloga("Timestep=%u, %d read m2d: {(%d,%d,%d),(%d,%d,%d)} from space\n",
+	uloga("Timestep=%u, %d read m2d: {(%llu,%llu,%llu),(%llu,%llu,%llu)} from space\n",
 		ts, rank_, xl,yl,zl, xu,yu,zu);
 #endif
 
@@ -172,12 +172,12 @@ static int couple_read_3d(unsigned int ts, int num_vars, enum transport_type typ
 	//get data from space
     set_offset_3d(rank_, npx_, npy_, npz_, spx_, spy_, spz_);
 	int elem_size = elem_size_;
-	int xl = offx_;
-	int yl = offy_;
-	int zl = offz_;
-	int xu = offx_ + spx_ - 1;
-	int yu = offy_ + spy_ - 1;
-	int zu = offz_ + spz_ - 1;
+	uint64_t xl = offx_;
+	uint64_t yl = offy_;
+	uint64_t zl = offz_;
+	uint64_t xu = offx_ + spx_ - 1;
+	uint64_t yu = offy_ + spy_ - 1;
+	uint64_t zu = offz_ + spz_ - 1;
 	double tm_st, tm_end, tm_max, tm_diff;
 	int root = 0;
 
@@ -228,7 +228,7 @@ static int couple_read_3d(unsigned int ts, int num_vars, enum transport_type typ
 }
 
 int test_get_run(enum transport_type type, int npapp, int npx, int npy, int npz,
-        int spx, int spy, int spz, int timestep, int dims, size_t elem_size,
+        uint64_t spx, uint64_t spy, uint64_t spz, int timestep, int dims, size_t elem_size,
         int num_vars, MPI_Comm gcomm)
 {
 	gcomm_ = gcomm;
